refactor: split progress reporter into modular package
- Create _reporting/ package with focused modules
- Split strategies, utils, and reporter into separate files
- Maintain backward compatibility via re-exports
- Split tests to match new structure (48 tests, up from 27)
