+++ tests/cli/eval/reporting/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the evaluation progress reporting module."""
